(feat): add jit for gausssian
